refactor: conform to snake_case filenames
PEP8: https://peps.python.org/pep-0008/#package-and-module-names

Mixing uppercase and lowercase in filenames can cause cross-platform issues
