+++ api/src/data_providers/get_repository.py
@@ -1,8 +1,8 @@
 from pymongo.mongo_client import MongoClient
 
 from config import config
-from data_providers.clients.mongodb.MongoDatabaseClient import MongoDatabaseClient
-from data_providers.repositories.TodoRepository import TodoRepository
+from data_providers.clients.mongodb.mongo_database_client import MongoDatabaseClient
+from data_providers.repositories.todo_repository import TodoRepository
 
 MONGO_CLIENT: MongoClient = MongoClient(
     host=config.MONGODB_HOSTNAME,
