+++ api/src/data_providers/repository_interfaces/todo_repository_interface.py
@@ -1,6 +1,6 @@
 from abc import ABCMeta, abstractmethod
 
-from entities.TodoItem import TodoItem
+from entities.todo_item import TodoItem
 
 
 class TodoRepositoryInterface(metaclass=ABCMeta):
