+++ api/src/features/todo/use_cases/add_todo.py
@@ -2,10 +2,10 @@ import uuid
 
 from pydantic import BaseModel, Field
 
-from data_providers.repository_interfaces.TodoRepositoryInterface import (
+from data_providers.repository_interfaces.todo_repository_interface import (
     TodoRepositoryInterface,
 )
-from entities.TodoItem import TodoItem
+from entities.todo_item import TodoItem
 
 
 class AddTodoRequest(BaseModel):
