+++ api/src/features/todo/use_cases/get_todo_by_id.py
@@ -3,10 +3,10 @@ from typing import cast
 from pydantic import BaseModel
 
 from common.exceptions import MissingPrivilegeException
-from data_providers.repository_interfaces.TodoRepositoryInterface import (
+from data_providers.repository_interfaces.todo_repository_interface import (
     TodoRepositoryInterface,
 )
-from entities.TodoItem import TodoItem
+from entities.todo_item import TodoItem
 
 
 class GetTodoByIdResponse(BaseModel):
