+++ api/src/features/todo/use_cases/update_todo.py
@@ -1,10 +1,10 @@
 from pydantic import BaseModel, Field
 
 from common.exceptions import MissingPrivilegeException
-from data_providers.repository_interfaces.TodoRepositoryInterface import (
+from data_providers.repository_interfaces.todo_repository_interface import (
     TodoRepositoryInterface,
 )
-from entities.TodoItem import TodoItem
+from entities.todo_item import TodoItem
 
 
 class UpdateTodoRequest(BaseModel):
