+++ api/src/tests/unit/data_providers/clients/mongodb/test_mongo_database_client.py
@@ -1,7 +1,7 @@
 import pytest
 
 from common.exceptions import NotFoundException, ValidationException
-from data_providers.clients.mongodb.MongoDatabaseClient import MongoDatabaseClient
+from data_providers.clients.mongodb.mongo_database_client import MongoDatabaseClient
 
 
 class TestMongoDatabaseClient:
