+++ api/src/tests/unit/features/todo/use_cases/test_delete_todo_by_id.py
@@ -1,7 +1,7 @@
 import pytest as pytest
 
 from common.exceptions import NotFoundException
-from data_providers.repository_interfaces.TodoRepositoryInterface import (
+from data_providers.repository_interfaces.todo_repository_interface import (
     TodoRepositoryInterface,
 )
 from features.todo.use_cases.delete_todo_by_id import (
