+++ documentation/docs/contribute/development-guide/coding/extending-the-api/adding-data-providers/02-repository-interfaces.md
@@ -5,7 +5,7 @@ A repository interface describes the incoming parameters and the type of the obj
 ```mdx-code-block
 import CodeBlock from '@theme/CodeBlock';
 
-import TodoRepositoryInterface from '!!raw-loader!@site/../api/src/data_providers/repository_interfaces/TodoRepositoryInterface.py';
+import TodoRepositoryInterface from '!!raw-loader!@site/../api/src/data_providers/repository_interfaces/todo_repository_interface.py';
 
 <CodeBlock language="jsx">{TodoRepositoryInterface}</CodeBlock>
 ```
